authenticate journals

+++ src/main/java/com/hammy/journalApp/service/JournalEntryService.java
@@ -33,7 +33,7 @@ public class JournalEntryService {
             journalEntry.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntry); // saved in local variable
             user.getJournalEntries().add(saved); // added in journal entries list
-            userService.saveEntry(user); // saved user in database with new journal entry
+            userService.saveUser(user); // saved user in database with new journal entry
         } catch (Exception e) {
             log.error("Exception ", e);
             throw new RuntimeException("An error has occurred while saving the entry.", e);
@@ -57,11 +57,22 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+
+            if(removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry.", e);
+        }
+        return removed;
     }
 
 }
